requirements.txt
discord.py==2.3.1
numpy>=1.26
pandas>=2.2.2
python-dotenv==1.0.0
PyYAML
pytest
requests==2.31.0
pydantic
datamodel-code-generator
uproot